+++ src/main/Reservas/Factura.java
@@ -2,26 +2,28 @@ package main.Reservas;
 
 public class Factura {
     private int nroFactura;
-    private int importe;
+    private Double importe;
 
-
-    //constructor
-    public Factura(int nroFactura, int importe) {
+    // constructor
+    public Factura(int nroFactura, Double importe) {
         this.nroFactura = nroFactura;
         this.importe = importe;
     }
 
-    //getters y setters
+    // getters y setters
     public int getNroFactura() {
         return nroFactura;
     }
+
     public void setNroFactura(int nroFactura) {
         this.nroFactura = nroFactura;
     }
-    public int getImporte() {
+
+    public Double getImporte() {
         return importe;
     }
-    public void setImporte(int importe) {
+
+    public void setImporte(Double importe) {
         this.importe = importe;
     }
